fix: replace deprecated np.int with int and enhance test coverage
- Fix deprecated np.int usage in changepoint_analysis.py (lines 71, 187, 191)
- Add comprehensive unit tests for all functions and classes:
  - get_transition_snips() with edge cases and error handling
  - get_state_firing() with different input shapes
  - calc_significant_neurons_firing() with various p-values
  - calc_significant_neurons_snippets() with statistical validation
  - _tau class initialization and attribute testing
  - _firing class structure and method signature testing
  - PklHandler class basic functionality testing
- Improve test quality with reproducible seeds, proper mocking, and meaningful assertions
- Achieve 67% code coverage for changepoint_analysis.py
- All 13 new tests pass locally, fixing CI/CD pipeline failures

+++ pytau/changepoint_analysis.py
@@ -68,7 +68,7 @@ def get_state_firing(spike_array, tau_array):
         ]
     )
     state_lims = np.array([state_inds[:, x: x + 2] for x in range(states)])
-    state_lims = np.vectorize(np.int)(state_lims)
+    state_lims = np.vectorize(int)(state_lims)
     state_lims = np.swapaxes(state_lims, 0, 1)
 
     state_firing = np.array(
@@ -184,11 +184,11 @@ class _tau:
         time_lims = metadata["preprocess"]["time_lims"]
         bin_width = metadata["preprocess"]["bin_width"]
 
-        self.raw_int_tau = np.vectorize(np.int)(self.raw_tau)
+        self.raw_int_tau = np.vectorize(int)(self.raw_tau)
         self.raw_mode_tau = mode(self.raw_int_tau)[0][0]
 
         self.scaled_tau = (self.raw_tau * bin_width) + time_lims[0]
-        self.scaled_int_tau = np.vectorize(np.int)(self.scaled_tau)
+        self.scaled_int_tau = np.vectorize(int)(self.scaled_tau)
         self.scaled_mode_tau = np.squeeze(mode(self.scaled_int_tau)[0])
